chore: rename BitbucketRepoBranchPicker folder to RepoBranchPicker
